Better handling of batch changes

+++ create.js
@@ -1,8 +1,10 @@
 'use strict';
 
 var validFunction      = require('es5-ext/function/valid-function')
+  , assign             = require('es5-ext/object/assign')
   , setPrototypeOf     = require('es5-ext/object/set-prototype-of')
   , d                  = require('d/d')
+  , lazy               = require('d/lazy')
   , ee                 = require('event-emitter')
   , memoize            = require('memoizee/lib/regular')
   , validSet           = require('es6-set/valid-set')
@@ -28,11 +30,16 @@ module.exports = memoize(function (Constructor) {
 	clear = Constructor.prototype.clear;
 	del = Constructor.prototype.delete;
 
-	Observable.prototype = ee(Object.create(Constructor.prototype, {
+	Observable.prototype = ee(Object.create(Constructor.prototype, assign({
 		constructor: d(Observable),
 		add: d(function (value) {
 			if (this.has(value)) return this;
 			add.call(this, value);
+			if (this.__onHold__) {
+				if (!this.__added__) this.__added__ = [value];
+				else this.__added__.push(value);
+				return this;
+			}
 			this.emit('change', { type: 'add', value: value });
 			return this;
 		}),
@@ -45,12 +52,42 @@ module.exports = memoize(function (Constructor) {
 		$clear: d(clear),
 		delete: d(function (value) {
 			if (!del.call(this, value)) return false;
+			if (this.__onHold__) {
+				if (!this.__deleted__) this.__deleted__ = [value];
+				else this.__deleted__.push(value);
+				return this;
+			}
 			this.emit('change', { type: 'delete', value: value });
 			return true;
 		}),
-		$delete: d(del)
-	}));
+		$delete: d(del),
+		_release_: d(function () {
+			var event, added = this.__added__, deleted = this.__deleted__;
+			if (added && added.length) {
+				if (deleted && deleted.length) {
+					event = { type: 'batch', added: added, deleted: deleted };
+				} else if (added.length === 1) {
+					event = { type: 'add', value: added[0] };
+				} else {
+					event = { type: 'batch', added: added };
+				}
+			} else if (deleted && deleted.length) {
+				if (deleted.length === 1) {
+					event = { type: 'delete', value: deleted[0] };
+				} else {
+					event = { type: 'batch', deleted: deleted };
+				}
+			}
+			this.__added__ = this.__deleted__ = this.__onHold__ = null;
+			if (!event) return;
+			this.emit('change', event);
+		})
+	}, lazy({
+		__onHold__: d('w', null),
+		__added__: d('w', null),
+		__deleted__: d('w', null)
+	}))));
 	defineProperty(Observable.prototype, isObservableSymbol, d('', true));
 
 	return Observable;
-});
+}));
